OSIS-724 Refactoring file structure
