+++ base/views/learning_units/proposal/__init__.py
@@ -1,0 +1,1 @@
+from base.views.learning_units.proposal import creation
