+++ base/views/learning_units/proposal/creation.py
@@ -31,7 +31,7 @@ from django.shortcuts import redirect, get_object_or_404
 from base import models as mdl_base
 from base.business.learning_unit import create_learning_unit, create_learning_unit_structure
 from base.business.learning_units.proposal.creation import create_proposal
-from base.forms.proposal.creation import LearningUnitProposalCreationForm, LearningUnitProposalForm
+from base.forms.learning_unit.proposal.creation import LearningUnitProposalCreationForm, LearningUnitProposalForm
 from base.models.enums import learning_unit_year_subtypes
 from base.models.learning_container import LearningContainer
 from base.models.person import Person
