--- base/views/learning_units/proposals/__init__.py
@@ -1,1 +1,0 @@
-from base.views.learning_units.proposals import creation
